Get CS converter working (#1395)
This change fixes a set of bugs that prevent the partially implemented Contrast Security converter from running to completion. Now that these bugs are fixed, we have a working converter to which we can add new rule conversions one at a time.

The bugs are all failures to check whether a rule's `DefaultConfiguration` is non-`null` before dereferencing it to obtain its `Level` property. If `DefaultConfiguration` is missing, `Level` should default to `FailureLevel.Warning`. So wrap up the null-checking and defaulting logic in a new helper method `FailureLevel GetRuleFailureLevel(string ruleId)`, and call it everywhere (including `ConstructNotImplementedRuleResult`).

Also:
- Fix some nits I ran across while studying the code.