Fix #1514: location.logicalLocation convenience setter mishandles null (#1515)
If you set `location.LogicalLocation` to `null`, the convenience property sets `location.LogicalLocations` to an array with a single `null` element, which serializes as `"logicalLocations": [ {} ]`. The right answer is for it to set `location.LogicalLocations` to `null`, which would be omitted from the serialization.

Add unit tests for the convenience property.