Fix #1680 (regex perf in WebRequest.TryParse) by avoiding backtracking. (#1612)
The bug was exposed by an actual web request which left the `=` sign off of the last in a long list of query parameters:
```
/getSomething?firstName=John&lastName=Smith&...&Address2=&Address3
```
Because our regex for query parameters expects an `=` after each query parameter name, the lack of the final `=` caused the engine to backtrack all the way to the beginning. By turning off backtracking within the individual parameter names and values, and by eliminating an unnecessary alternation, we minimize the number of states the engine has to save.