Fix #1532: Multitool transform mishandles dottedQuadFileVersion.
The SARIF v1 spec does not specify a format for tool.fileVersion (although on the Windows platform, the spec anticipated that it would be of the form n.n.n.n where each n is one or more decimal digits).

The SARIF v2 spec is more prescriptive: it defines a property toolComponent.dottedQuadFileVersion with exactly that format. So if the v1 file contained a tool.fileVersion in any other format, the best we can do is put it in the property bag.