Fix #1485: SARIF1015 mishandles run.originalUriBaseIds (@kupsch) (#1644)
SARIF1015 complained if a `uri` property in one of the `artifactLocation` objects in `run.originalUriBaseIds` was relative. But that is allowed as long as `artifactLocation.uriBaseId` is also present.

The actual requirement is:

1. If `uriBaseId` is absent, then _either_ `uri` must be absolute or it must be absent.
2. If `uriBaseId` is present, the `uri` must be relative.

But <span>#</span>2 is true for _all_ `artifactLocation` objects, not just those in `run.originalUriBaseIds`. Therefore, in this PR:

- We remove the logic dealing with `run.originalUriBaseIds` from SARIF1015.
- We introduce a new rule SARIF1018 that enforces <span>#</span>1 in `run.originalUriBaseIds`.
- We file issue #1643, "New rule: If artifactLocation.uriBaseId is present, uri must be relative" to enforce <span>#</span>2 for all `artifactLocation` objects. But that is beyond the scope of this PR.

Also:
- DRY out the construction of the names of the validation rule test files.
- Upgrade the test files for SARIF1015 (which are affected by this change) to the final SARIF version, to make it easier to reliably make the required changes in `run.originalUriBaseIds`.

NOTE: #1485 also reported that the web site's validation page did not accept files with the (valid) filename extension .sarif.json. I filed microsoft/sarif-website#104 for that, and labeled it `m156` because _this_ bug is `m156`. I'll do that one next.