Fix #1592: Merge command returns invalid SARIF if there are 0 input files. (#1648)
The product code fix is straightforward. To understand the _reason_ for the change, see the comment thread in #1592.

The test code looks similar to code you've seen before, but there is a subtlety, explained in a big WARNING comment. I don't think it's necessary to address this issue now, and possibly ever.