Fix #1629: Page command shouldn't throw on options error. (#1654)
Some option validations (for example, "--index must be non-negative") can be performed without reading the input SARIF file. When such a validation fails, we can simply print an error message and exit. By not throwing an exception in these cases, we save the user from having to dig the error message out of an exception dump.

On the other hand, other option validations (for example, "--run-index must be less than the number of runs in the input SARIF file") can only be performed once we've started reading the input file. When such a validation fails, we have to throw an exception. This PR does not attempt to improve the experience in that case.