Fix #1577: PrereleaseCompatibilityTransformer mishandles date/times in property bags. (#1651)
The `PrereleaseCompatibilityTransformer` was converting SARIF like this:
```
"properties": {
  "alertDiscoveryDate": "2019-06-22T19:06:31.209Z"
}
```
to this invalid SARIF:
```
"properties": {
  "alertDiscoveryDate": 6/22/2019 7:06:31 PM
}
```
Note that the date format has changed, but worse, it's not enclosed in quotes.

The fix is explained in a comment on the new method `SarifUtilities.DeserializeObject<T>`.

The fix then exposed a subtle bug in our `DateTimeConverter` class that caused several unit tests to fail. The fix for _that_ bug is explained in a comment in `DateTimeConverter.WriteJson`.