Fix #1656: match-results-forward fails in RemapIndicesVisitor (#1663)
This change fixes the treatment of logical locations in the `RemapIndicesVisitor`.

As the code stood, if a result implicated a nested logical location (and most logical locations _are_ nested, for example `Namespace.Class.Method`), the visitor would populate `run.logicalLocations` with only that logical location, omitting its ancestors (`Namespace.Class` and `Namespace`). Worse yet, the visitor left the `logicalLocation` object's `parentIndex` property alone, so it pointed to whatever array index the parent had occupied in the original run (that is, in the baseline or current run, as opposed to the merged run). This led to index out of range errors in the `SarifLogResultMatcher`.

The visitor treated nested artifacts correctly, copying all the ancestors from the original run, and adjusting both `artifact.parentIndex` and `artifact.location.index` correctly. We rewrote the treatment of logical locations to mirror the treatment of artifacts.

Also:
- Fix #1630: Multitool page command --force defaults to true and can't be changed.
- Make success/failure symbols ALL_CAPS and internal.
- Rename "file" to "artifact" throughout.