Improve handling of DateTime property bag properties
Fix a bug in PropertyBagConverter that prevented `DateTime`-valued property bag properties from round-tripping correctly. With this fix in place, it is no longer necessary to work around the problem by circumventing Newtonsoft.Json's handling of `DateTime`s, and so the helper method `SarifUtilities.DeserializeObject` is no longer needed.

Revert calls to the deleted method to their state before the method was introduced. In most cases, the previous call was to `JsonConvert.DeserializeObject`. In one case, the previous call was to `JObject.Parse`.