Result filtering, part 1 (#1708)
In this PR we partially address the second of Scott's four log partitioning scenarios:

```
// Build a filtered log including all results the predicate returns ‘true’ for.
// Should enumerate the Results only once.
SarifLog Filter(SarifLog log, Func<Result, ResultIndex, bool> includePredicate);
```

I chose to do this one first because it seemed an easy one to start with.

This is a partial implementation, demonstrating that I can filter `run.artifacts` along with `run.results`. Still TODO:

- Enhance the test file to demonstrate how the `FilteringVisitor` picks up the other `result`-scoped `artifactLocations` (for example, in `result.relatedLocations`).
- Enhance the test file to demonstrate how the visitor picks up `run`-scoped `artifactLocations` (for example, in `run.attachments[].location`).

In email @michaelcfanning suggested doing the `run`-scoped properties separately, but the single call to `FilteringVisitor.VisitArtifactLocation` works for everything, so no additional product code is needed to handle those properties.

In Scott's proposed API, the predicate took a `ResultIndex` (run index/result index pair) in addition to the `Result` itself. After discussion, Scott and I agreed that this parameter isn't necessary in this scenario, and might possibly never be needed, so we didn't include it.

Also differing from Scott's proposal, I defined a delegate type `FilteringVisitor.FilteringPredicate` rather than using a `Func`. I'm not sure I like it, partly because I can't decide whether that delegate type belongs in the `FilteringVisitor` (which means that callers of the wrapper function `SarifPartitioner.Filter` have to know that the `FilteringVisitor` exists, which breaks encapsulation) or in `SarifPartitioner` (which means that `FilteringVisitor` has to know that the wrapper `SarifPartitioner` exists, which seems inside-out).

Finally, I wanted to use our `FileDiffingFunctionalTest` infrastructure, but it didn't meet my needs because I need to run the filtering algorithm over the same file repeatedly, but with different filtering predicates. There was no way to specify a parameter when overriding `FileDiffingUnitTests.ConstructTestOutputFromInputResource`. So I changed the signature to accept a parameter; then I had to modify the other classes derived from `FileDiffingUnitTests` to declare the parameter (even though they don't use it).