Fix #1737: SarifLogBaseliner does not populate firstDetectionTimeUtc
Previously, the code just asked whether or not a result from the current log matched a result in the baseline. Now, we actually extract the matching result from the baseline so we can roll its first detection time forward. If the matching result has no first detection time, we fall back to the time that the run was performed. And if we can't find that, we fall back to "now" as the first detection time.