Response to comments
Some formatting.
Removes seek in ctor.
Changes JsonPositionedTextReader(Stream) to be a factory method so we can enforce that the stream provided is seekable.
Adds some unit tests for DelegatingStream.
Moves DelegatingStream next to JsonPositionedTextReader since it is a helper class.
Adds test cases for deferred collections reading from the new stream API.