Consolidate PR validation into one file with multiple jobs (#2831)
Both build-and-test and build-multitool-for-npm had jobs that were named "run". It turns out that job names must be globally unique. In some contexts, including branch protection and on the mobile github app, only the job name shows. So, you would see "Run" and not "Build and Test / Run" and "Build Multitool for npm / Run". Worse, you could not require both of these jobs to pass in branch protection.

This change puts all of the required PR validation jobs together in one file to guarantee that future job names are unique and to encourage future jobs to be given equally good names.